14/06/2021

+++ app/Http/Controllers/EmpleadoController.php
@@ -51,7 +51,9 @@ class EmpleadoController extends Controller
 
         Empleado::insert($datosEmpleado);
 
-        return response()->json($datosEmpleado);
+        //return response()->json($datosEmpleado);
+
+        return redirect('empleado')->with('mensaje','Empleado Agregado con exito');
     }
 
     /**
@@ -111,7 +113,13 @@ class EmpleadoController extends Controller
     public function destroy($id)
     {
         //
-        Empleado::destroy($id);
-        return redirect('empleado');
+        $empleado=Empleado::findOrFail($id);
+        if(Storage::delete('public/'.$empleado->Foto)){
+            Empleado::destroy($id);
+
+        }
+       
+        //return redirect('empleado');
+        return redirect('empleado')->with('mensaje','Empleado Borrado');
     }
 }
